+++ projects/ruby-3.3.10/ext/openssl/ossl_bn.c
@@ -385,7 +385,7 @@ ossl_bn_to_i(VALUE self)
     if (!(txt = BN_bn2hex(bn))) {
 	ossl_raise(eBNError, NULL);
     }
-    num = rb_cstr_to_inum(txt, 16, Qtrue);
+    num = rb_cstr_to_inum(txt, 16, (int)Qtrue);
     OPENSSL_free(txt);
 
     return num;
